+++ src/Libraries/RevitNodes/PerformanceAdviser/FailureMessage.cs
@@ -2,9 +2,9 @@
 using Autodesk.DesignScript.Runtime;
 using RevitServices.Persistence;
 using System.Collections.Generic;
+using Revit.Elements;
 
-
-namespace Revit.Elements
+namespace Revit.PerformanceAdviser
 {
     /// <summary>
     /// Performance Adviser Failure Message
@@ -37,12 +37,12 @@ namespace Revit.Elements
         /// <summary>
         /// The Failing Elements of the message.
         /// </summary>
-        public ICollection<Element> FailingElements
+        public ICollection<Revit.Elements.Element> FailingElements
         {
             get
             {
                 var failingIds = InternalElement.GetFailingElements();
-                var failingElements = new List<Element>();
+                var failingElements = new List<Revit.Elements.Element>();
 
                 foreach (var failingId in failingIds)
                 {
@@ -68,27 +68,7 @@ namespace Revit.Elements
                 return InternalElement.GetSeverity().ToString();
             }
         }
-     
-        /// <summary>
-        /// Post Messages in Document
-        /// </summary>
-        /// <param name="messages"></param>
-        public static void PostInDocument(IEnumerable<FailureMessage> messages)
-        {
-            
-            var document = DocumentManager.Instance.CurrentDBDocument;
 
-            foreach (var msg in messages)
-            {
-                Autodesk.Revit.DB.Transaction tr = new Autodesk.Revit.DB.Transaction(document);
-                tr.Start("Reporting");
-
-                document.PostFailure(msg.InternalElement);
-
-                tr.Commit();
-            }
-
-        }
 
 
         public override string ToString()
