+++ src/Libraries/RevitNodes/PerformanceAdviser/PerformanceAdviserRule.cs
@@ -4,9 +4,9 @@ using Autodesk.DesignScript.Runtime;
 using Autodesk.Revit.DB;
 using RevitServices.Persistence;
 using RevitServices.Transactions;
+using Revit.Elements;
 
-
-namespace Revit.Elements
+namespace Revit.PerformanceAdviser
 {
     /// <summary>
     /// Performance Adviser Rule
@@ -31,7 +31,7 @@ namespace Revit.Elements
         {
             get
             {
-                PerformanceAdviser adviser = PerformanceAdviser.GetPerformanceAdviser();
+                Autodesk.Revit.DB.PerformanceAdviser adviser = Autodesk.Revit.DB.PerformanceAdviser.GetPerformanceAdviser();
                 return adviser.GetRuleName(this.InternalId);
             }
         }
@@ -43,7 +43,7 @@ namespace Revit.Elements
         {
             get
             {
-                PerformanceAdviser adviser = PerformanceAdviser.GetPerformanceAdviser();
+                Autodesk.Revit.DB.PerformanceAdviser adviser = Autodesk.Revit.DB.PerformanceAdviser.GetPerformanceAdviser();
                 return adviser.GetRuleDescription(this.InternalId);
             }
         }
@@ -55,7 +55,7 @@ namespace Revit.Elements
         {
             get
             {
-                PerformanceAdviser adviser = PerformanceAdviser.GetPerformanceAdviser();
+                Autodesk.Revit.DB.PerformanceAdviser adviser = Autodesk.Revit.DB.PerformanceAdviser.GetPerformanceAdviser();
                 // Check if the rule is activated
                 return adviser.IsRuleEnabled(this.InternalId);
             }
@@ -68,7 +68,7 @@ namespace Revit.Elements
         {
             set
             {
-                PerformanceAdviser adviser = PerformanceAdviser.GetPerformanceAdviser();
+                Autodesk.Revit.DB.PerformanceAdviser adviser = Autodesk.Revit.DB.PerformanceAdviser.GetPerformanceAdviser();
                 // Activate or deactivate the rule
                 adviser.SetRuleEnabled(this.InternalId, value);
             }
@@ -96,7 +96,7 @@ namespace Revit.Elements
             var document = DocumentManager.Instance.CurrentDBDocument;
             TransactionManager.Instance.EnsureInTransaction(document);
 
-            PerformanceAdviser adviser = PerformanceAdviser.GetPerformanceAdviser();
+            Autodesk.Revit.DB.PerformanceAdviser adviser = Autodesk.Revit.DB.PerformanceAdviser.GetPerformanceAdviser();
             IList<PerformanceAdviserRuleId> rulesIdsToExecute = new List<PerformanceAdviserRuleId>();
 
             foreach (var rule in rules) rulesIdsToExecute.Add(new PerformanceAdviserRuleId(rule.RuleId));
@@ -131,7 +131,7 @@ namespace Revit.Elements
         [SupressImportIntoVM]
         public override string ToString()
         {
-            PerformanceAdviser adviser = PerformanceAdviser.GetPerformanceAdviser();
+            Autodesk.Revit.DB.PerformanceAdviser adviser = Autodesk.Revit.DB.PerformanceAdviser.GetPerformanceAdviser();
             return string.Format("{0} : {1}", adviser.GetRuleName(this.InternalId), adviser.GetRuleDescription(this.InternalId));
         }
